StatEngine and StatReport development

+++ Next-Level Sonics City/Assets/Model/Statistics/StatReport.cs
@@ -6,15 +6,14 @@ namespace Model.Statistics
 		public int Quarter { get; internal set; }
 		public float Happiness { get; internal set; }
 
-		public int IncomeTax { get; internal set; }
-		public int ResidenceTax { get; internal set; }
-		public int DestroyIncomes { get; internal set; }
-		public int BuildExpenses { get; internal set; }
-		public int MaintainanceCosts { get; internal set; }
-
-		public int Incomes { get; internal set; }
-		public int Expenses { get; internal set; }
-		public int Profit { get; internal set; }
+		public float IncomeTax { get; internal set; }
+		public float ResidenceTax { get; internal set; }
+		public float DestroyIncomes { get; internal set; }
+		public float BuildExpenses { get; internal set; }
+		public float MaintainanceCosts { get; internal set; }
+		public float Incomes { get; internal set; }
+		public float Expenses { get; internal set; }
+		public float Profit { get; internal set; }
 
 		public int Population { get; internal set; }
 		public int PopulationChange { get; internal set; }
